Work with GitHub repos by default
