--- src/common/generatorType.ts
@@ -1,15 +1,0 @@
-import type { Repo } from "~common/repo";
-
-const defaultGenerator = (repo: Repo): WorkflowSteps => {
-  return []
-}
-
-type StepsGeneratorFn = typeof defaultGenerator
-
-export type StepsGenerator = {
-  name: string;
-  fn: StepsGeneratorFn;
-}
-
-export type WorkflowSteps = object[]
-
