+++ src/console.ts
@@ -1,4 +1,4 @@
-import { GitHubRepo, REPO_TYPES } from "~common/repo";
+import { Repo } from "~common/repo";
 import { genWorkflows, validateRepo } from "~lib/candidate";
 
 const repos: string[] = [
@@ -7,8 +7,8 @@ const repos: string[] = [
 
 for (let i = 0; i < repos.length; i++) {
   const repoName: string = repos[i];
-  const repo = new GitHubRepo(repoName);
-  console.log("Validating: ", repoName, " result: ", validateRepo(repoName, REPO_TYPES.GITHUB_REPO));
+  const repo = new Repo(repoName);
+  console.log("Validating: ", repoName, " result: ", validateRepo(repoName));
 
   console.log("Generating Workflows: ", repoName, " results: ");
   console.log(genWorkflows(repo));
