--- src/generators/github/index.ts
@@ -1,8 +1,0 @@
-import type { StepsGenerator } from "~common/generatorType";
-import NPMGenerator from "./npm"
-
-const generators: { [key: string]: StepsGenerator } = {
-  npm: NPMGenerator
-}
-
-export default generators;
