+++ src/lib/candidate.test.ts
@@ -1,10 +1,9 @@
 import { expect, test } from "@jest/globals"
-import { REPO_TYPES } from "~common/repo"
 
 import { validateRepo, genWorkflows } from "./candidate"
 
 test("validateRepo is not yet ready", async () => {
-  expect(validateRepo('test', REPO_TYPES.GITHUB_REPO)).toBe(false)
+  expect(validateRepo('test')).toBe(false)
 })
 
 test("genActions", async () => {
