+++ src/lib/candidate.ts
@@ -1,8 +1,7 @@
 #!/usr/bin/env node
-import type { FilterMatchHeuristic } from "~common/filterType";
-import { baseGitHubWorkflow } from "~common/githubType";
-import { REPO_TYPES, Repo } from "~common/repo";
-import { calcHeuristics } from "~filters";
+import { baseGitHubWorkflow } from "~common/github-base-workflow";
+import type { Repo } from "~common/repo";
+import { FilterMatchHeuristic, calcHeuristics } from "~filters";
 import { generateWorkflowSteps } from "~generators";
 
 export type WorkflowFile = {
@@ -10,7 +9,7 @@ export type WorkflowFile = {
   workflow: object
 }
 
-export const validateRepo = (name: string, type: REPO_TYPES): boolean => {
+export const validateRepo = (name: string): boolean => {
   // Evaluates if a repo is a candidate or not
   return false;
 }
@@ -18,31 +17,23 @@ export const validateRepo = (name: string, type: REPO_TYPES): boolean => {
 export const genWorkflows = (repo: Repo): WorkflowFile[] => {
   let toolNames = findBuildTools(repo);
   let workflows = [];
-  
-  // Only Github Repos are supported for now
-  switch (repo.source) {
-    case REPO_TYPES.GITHUB_REPO:
-      toolNames.forEach(toolName => {
-        const workflow = {
-          ...baseGitHubWorkflow,
-          jobs: {
-            ...baseGitHubWorkflow.jobs,
-            build: {
-              ...baseGitHubWorkflow.jobs.build,
-              steps: baseGitHubWorkflow.jobs.build.steps.concat(
-                generateWorkflowSteps(repo, toolName)
-              )
-            }
-          }
-        }
 
-        workflows.push({ name: `${toolName}_deploy`, workflow })
-      });
-      break;
-  
-    default:
-      break;
-  }
+  toolNames.forEach(toolName => {
+    const workflow = {
+      ...baseGitHubWorkflow,
+      jobs: {
+        ...baseGitHubWorkflow.jobs,
+        build: {
+          ...baseGitHubWorkflow.jobs.build,
+          steps: baseGitHubWorkflow.jobs.build.steps.concat(
+            generateWorkflowSteps(repo, toolName)
+          )
+        }
+      }
+    }
+
+    workflows.push({ name: `${toolName}-deploy`, workflow })
+  });
 
   return workflows;
 }
@@ -54,7 +45,7 @@ const findBuildTools = (repo: Repo): string[] => {
   const evaluations: FilterMatchHeuristic[] = calcHeuristics(repo);
 
   evaluations.forEach(heurictic => {
-    if (heurictic.score > 0) {
+    if (heurictic.dataPoints.length > 0) {
       // TODO: better filtering
       tools.push(heurictic.name);
     }
